A scenario is added to verify the error messages when entering email with wrong data in the login form
