+++ src/main/java/utils/BrowserManagement.java
@@ -52,6 +52,7 @@ public class BrowserManagement {
                 ChromeOptions chromeOptions = new ChromeOptions();
                 chromeOptions.addArguments("--start-maximized");
                 chromeOptions.addArguments("--incognito");
+                chromeOptions.setHeadless(true);
                 desiredCapabilities.setCapability(ChromeOptions.CAPABILITY, chromeOptions);
                 driver = new ChromeDriver(chromeOptions);
                 break;
